pyproject.toml
[tool.poetry]
name = "mybot"
version = "0.1.0"
description = "mybot"
authors = ["meetwq <meetwq@gmail.com>"]
license = "MIT"

[[tool.poetry.source]]
name = "tsinghua"
url = "https://mirrors.tuna.tsinghua.edu.cn/pypi/web/simple/"
default = true

[tool.poetry.dependencies]
python = "^3.7.3"
nonebot2 = "^2.0.0-beta.1"
nonebot-adapter-onebot = "^2.0.0-beta.1"
nonebot-plugin-apscheduler = "^0.1.2"
nonebot-plugin-manager = "^0.5.7"
nonebot-plugin-alias = "^0.2.2"
nonebot-plugin-withdraw = "^0.1.6"
nonebot-plugin-petpet = "^0.1.13"
nonebot-plugin-logo = "^0.1.1"
nonebot-plugin-memes = "^0.1.2"
nonebot-plugin-asoulcnki = "^0.1.3"
nonebot-plugin-shindan = "^0.1.8"
nonebot-plugin-htmlrender = ">=0.0.4.1"
aliyunpan = "^2.10.0"
baike = "^1.5.0"
cachetools = "^4.2.0"
cos-python-sdk-v5 = "^1.9.0"
feedparser = "^6.0.0"
httpx = ">=0.19.0"
imageio = "^2.13.0"
Jinja2 = "^3.0.0"
langid = "^1.1.0"
lxml = "^4.7.0"
mcstatus = "^6.5.0"
numpy = "^1.21.0"
Pillow = "^8.4.0"
PixivPy-Async = "^1.2.0"
playwright = "^1.17.0"
pydub = "^0.25.0"
python-Levenshtein-wheels = "^0.13.0"
tencentcloud-sdk-python = "^3.0.0"
thefuzz = "^0.19.0"
wolframalpha = "^5.0.0"

[tool.poetry.dev-dependencies]
nb-cli = "^0.6.0"
nonebot-plugin-test = "^0.3.0"
autopep8 = "^1.6.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
